Some JMA format and code improvements.
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@1861 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/jma/jma.h
@@ -46,8 +46,8 @@ namespace JMA
   
   struct jma_file_info : jma_file_info_base
   {
-    unsigned int date;
-    unsigned int time;
+    unsigned short date;
+    unsigned short time;
     const unsigned char *buffer;
   };
